renamed server files from snake-case to camelCase. Also updated all the font awesome icons to v5.
