+++ .gitignore
@@ -5,7 +5,7 @@ coverage
 .nyc_output
 npm-debug.log
 src/server/logs/
-src/front-end/public/javascripts/riot-tags.bundle.js
+src/front-end/public/javascripts/riotTags.bundle.js
 src/front-end/public/dist/*.css
 src/front-end/public/dist/*.js
 config/local.json
