+++ src/front-end/tags/loading-spinner/loading-spinner.tag
@@ -2,7 +2,7 @@
   <div show={ opts.loadingFlag } class="column has-text-centered has-text-grey">
         <p> { opts.text } 
           <span class="icon is-medium">
-            <i class="fa fa-cog fa-spin"></i>
+            <i class="fas fa-cog fa-spin"></i>
           </span>
         </p>
   </div>
