+++ src/front-end/views/contribute-components/overview.pug
@@ -114,4 +114,4 @@ block info
           |
           | The common goal we all share is to grow the community of contributors and make Mathbook the best online resource for learning mathematics. 
           span.icon
-            i.fa.fa-smile-o
+            i.far.fa-smile
