+++ src/front-end/views/contribute-components/review-tutorial.pug
@@ -37,7 +37,7 @@ block info
           | think about the stuff you are writing about and makes sure that you really know your stuff. It can also appear to be a little stressful because 
           | it feels as if the people reviewing your tutorial are judging your knowledge of a certain topic and subject, but that's ok, it's for the greater good.
           span.icon
-            i.fa.fa-smile-o
+            i.far.fa-smile
           br
           | All jokes aside, we would assume that people are reviewing your tutorial because they know a thing or two about it and they feel that they might learn
           | a thing or two from you. Everyone wins.
