+++ src/front-end/views/layout.pug
@@ -78,7 +78,7 @@ html
               strong Mathbook
             |  would not be possible without its talented Contributors.
             a.icon.my-footer-text(target='_blank', href='https://github.com/mathbook-io/mathbook/contributors.md')
-              i.fa.fa-github
+              i.fab.fa-github
           p.my-footer-text
             | Licensed under 
             a.my-footer-text(target='_blank', href='https://github.com/mathbook-io/mathbook/blob/develop/LICENSE')
