+++ src/front-end/views/login.pug
@@ -8,7 +8,7 @@ block content
       a.button.is-white(href='login/authenticate')
         span Login with GitHub 
         span.icon
-          i.fa.fa-github
+          i.fab.fa-github
   section.section
     .container
       .content
