+++ src/server/apis/createTutorial.js
@@ -1,7 +1,7 @@
 "use strict"
 const Promise = require("bluebird")
 const _ = require("lodash")
-const github = require("../github-client")
+const github = require("../githubClient")
 const Base64 = require("js-base64").Base64
 const errors = require("../errors")
 const transformError = require("../transformers/errorTransformer")
