+++ src/server/apis/saveTutorial.js
@@ -3,7 +3,7 @@
 const Promise = require("bluebird")
 const moment = require("moment")
 const _ = require("lodash")
-const github = require("../github-client")
+const github = require("../githubClient")
 const Base64 = require("js-base64").Base64
 const errors = require("../errors")
 const transformError = require("../transformers/errorTransformer")
