+++ src/server/apis/submitTutorial.js
@@ -2,7 +2,7 @@
 const config = require("../../../config/config")()
 const domain = config.get("bin.domain")
 const _ = require("lodash")
-const github = require("../github-client")
+const github = require("../githubClient")
 const errors = require("../errors")
 const transformError = require("../transformers/errorTransformer")
 const constants = require("../../../config/constants.json")
